develop-be-django with Custom User Model and Simple JWT/requirements.txt
django == 3.2.8
djangorestframework == 3.12.4
djangorestframework_simplejwt
psycopg2
django-simple-history
django-heroku
gunicorn
